Combined files containing main methods into one main adjacent file and renamed main.py to match naming convention.
